+++ topic_selector_lru.go
@@ -3,7 +3,7 @@ package mercure
 import (
 	"hash/fnv"
 
-	"github.com/hashicorp/golang-lru"
+	lru "github.com/hashicorp/golang-lru"
 )
 
 // Gather stats to find the best default values.
@@ -12,7 +12,7 @@ const (
 	DefaultTopicSelectorStoreLRUShardCount         = int64(256) // 2.5 million entries.
 )
 
-// NewTopicSelectorStoreLRU creates a TopicSelectorStore with an LRU cache
+// NewTopicSelectorStoreLRU creates a TopicSelectorStore with an LRU cache.
 func NewTopicSelectorStoreLRU(maxEntriesPerShard, shardCount int64) (*TopicSelectorStore, error) {
 	if maxEntriesPerShard == 0 {
 		return &TopicSelectorStore{}, nil
@@ -20,6 +20,7 @@ func NewTopicSelectorStoreLRU(maxEntriesPerShard, shardCount int64) (*TopicSelec
 	if shardCount == 0 {
 		shardCount = DefaultTopicSelectorStoreLRUShardCount
 	}
+
 	lruMap := make(shardedLRUCache, shardCount)
 	for i := 0; i < int(shardCount); i++ {
 		lruMap[i], _ = lru.New(int(maxEntriesPerShard))
@@ -36,11 +37,13 @@ func (c *shardedLRUCache) Get(k interface{}) (interface{}, bool) {
 
 func (c *shardedLRUCache) Set(k interface{}, v interface{}, _ int64) bool {
 	c.getShard(k).Add(k, v)
+
 	return true
 }
 
 func (c *shardedLRUCache) getShard(k interface{}) *lru.Cache {
 	h := fnv.New32a()
 	h.Write([]byte(k.(string)))
+
 	return (*c)[int(h.Sum32())%len(*c)]
 }
